t/cmd/lfstest-count-tests.go: close lock file
In commit 1926e942971a237c4c34e67e30fcd8a4b1d8244e of PR #3125 we
migrated our test suite to use the "prove" command, and introduced
the lfstest-count-tests utility to help control a common instance of
our lfstest-gitserver program, which can be shared and reused by all
the individual test scripts.  In that PR we also moved our test
suite into the "t" directory, and added a Makefile with a "test"
target that runs the "prove" command over all the individual test
scripts; this is still the framework we use today.

The lfstest-count-tests utility did not work as intended on the
Windows platform, however, as noted in the description of commit
9b73c8096ac546de0a1b076453ec548b1c5bc7ae in the same PR:

  On the AppVeyor CI Windows environment, there exist issues that prevent
  the lfstest-count-tests program from acquiring the test_count.lock file
  correctly.

Rather than debug this problem, several workarounds were added, such
as the GIT_LFS_NO_TEST_COUNT environment variable, which disables the
use of the lfstest-count-tests utility in our individual t/t-*.sh
test scripts, and the GIT_LFS_LOCK_ACQUIRE_DISABLED environment variable,
which is used to prevent the lfstest-count-tests utility from trying
to acquire the test_count.lock file at all on Windows.

The bug which causes the lfstest-count-tests program to be unable to
acquire the lock file the second time the program is run is the result
of the lock file not being removed when the program runs the first time.
This in turn is due to the fact that the file handle returned by
the call to the OpenFile() function of the "os" package is never
released with a call to Close(), so the subsequent call to Remove()
results in failure.

On Unix systems, this does not occur, because the file can be unlinked
while file handles to it remain open.  On Windows, however, all file
handles must be closed before the file can be deleted.

We fix the underlying bug here by simply calling Close() on the new
file handle immediately after creating the file, so long as no error
was returned by the call to OpenFile().

In addition, we make a number of other fixes and enhancements to the
lfstest-count-tests program.  We use the more conventional permissions
mode of 0777 for any directories we create, rather than 0666.
We include the O_RDWR flag with the others we pass to the OpenFile()
function, since we are technically supposed to provide at least one
of O_RDWR, O_RDONLY, and O_WRONLY.  And we output an error message
if the deferred call to our release() function is unable to delete
the lock file.