script/build-git: use custom libcurl on macOS
In commit b9f7f5ff88a5ff776a2420365c90bff5f9d69723 of PR #5813 we
updated our script/build-git script to install the version of Git
it builds into the location provided by the GIT_INSTALL_DIR
environment variable, if one is set; otherwise, /usr/local is used
as a default.

At the same time, we also changed the GitHub Actions workflow for
our CI jobs so that on macOS, when running our build-latest and
build-earliest jobs, we set the GIT_INSTALL_DIR environment variable
to a path of our choosing, and then also place that path at the start
of the list of paths in the PATH environment variable.  This ensures
that on macOS Actions runners, our custom version of Git is used in
preference to the one pre-installed on the runners by Homebrew.

However, in our build-earliest jobs, our custom build of Git is
still linked against the libcurl dynamic library installed by macOS
in /usr/lib, rather than the one our script/build-git script installs
and links using Homebrew.  (Technically, macOS Actions runners now
come with the Homebrew version installed but not linked, so our
script actually only performs the latter step.)

The reason our build of Git is linked against the default macOS version of
libcurl in our build-earliest CI jobs is that the Makefile used to build
Git, prior to commit git/git@23c4bbe28e61974577164db09cbd1d1c7e568ca4 of
Git version 2.20.0, does not use the output of "curl-config --libs" to
populate its CURL_LIBCURL variable but just adds the -lcurl argument
directly.  (Note that if the CURLDIR variable is defined in the build
configurations settings, its value takes precedence over these
defaults.) The CURL_LIBCURL variable is then used when building certain
programs like git-http-push(1), so these end up linked against the
macOS version of libcurl in /usr/lib, rather than the one we have
installed with Homebrew.

Until recently this has not caused problems, but at the moment, the
8.7.1 version of libcurl currently shipped with macOS 13 and 14 (Ventura
and Sonoma) has a regression which affects the programs used by
git-http-backend(1), including git-http-push, as detailed in
curl/curl#13229.

The consequence is that some of our tests fail because they happen to
encounter this regression.  In particular, the "does not look in current
directory for git with credential helper" test in our t/t-path.sh
test script commits a compiled Go binary file to a Git repository
and tries to push it, and when our lfstest-gitserver program passes
the HTTP POST request body on the git-http-backend program, the large
binary data in the request body includes byte sequences which trigger
the bug, resulting in a Git push failure and a test failure.

To resolve this problem and avoid similar ones in the future, we
revise our script/build-git script to set the CURLDIR variable
with the output of "curl-config --prefix" when running on macOS,
and then include this setting in the Makefile configuration we use
to build our custom Git.

As we have already run "brew link" to ensure "curl-config" will be
located in the /opt/homebrew/bin directory, which by default is ahead
of /usr/bin in the list of paths in PATH on our macOS Actions runners,
this means the Homebrew-installed "curl-config" program will be found,
and so we will then always link its dynamic library into the Git
programs we build on macOS.

These changes have no effect in our build-latest CI jobs, because
Git's Makefile already uses "curl-config" to set its CURL_LIBCURL
variable, but will resolve the problems we see at the moment in
our build-earliest CI jobs on macOS.