t/cmd/lfstest-gitserver.go: fix default URL files
Our lfstest-gitserver test utility program was updated in commit
daba49ae43340ee30b2e7c9a6b8d5927dbd71f51 of PR #1893 to help test
Git LFS's support for the use of client TLS/SSL certificates, and as
part of this change, the program was adjusted to write out several
new files upon startup with the values of the client certificate,
key, and URL to be used when testing this Git LFS feature.

In practice, the names of these files are always supplied by our
test scripts in LFSTEST_CLIENT_* environment variables, as defined
in our t/testenv.sh script and then passed to our lfstest-count-tests
test utility program in the setup() shell function that is run at the
start of every test script.  The lfstest-count-tests program then
invokes the lfstest-gitserver program, when necessary, and so the
lfstest-gitserver program inherits the LFSTEST_CLIENT_* environment
variables and uses the filenames defined by them.

However, should the lfstest-gitserver program be started manually,
without all of these environment variables, it will use a set of
default filenames for the files it creates after starting.  Two
of these default filenames are identical, specifically the default
names of the files containing the URLs the tests should use to contact
the test server when validating regular TLS/SSL behaviour and TLS/SSL
behaviour with a client certificate.  As such, the URL for the former
would be unavailable, as the file containing it would be overwritten
by the file containing the latter's URL.

We correct this minor issue now by defining a unique default name for
the file which contains the URL the tests should use when validating
TLS/SSL behaviour with a client certificate.  Note again, though, that
this makes no difference in the context of our test suite as it always
supplies unique, non-default values for all these files' names in its
environment variables.