t/t-clone.sh: unused TLS client certficate cleanup
In commit c54c9da0735297c22cf52c1258e711fa76357686 of PR #5657 we
added support for the use of client TLS/SSL certificate and key files
in a user's home directory, specifically, when their paths are
prefixed with a tilde character.  This change aligned Git LFS's
support of the "http.<url>.sslCert" and "http.<url>.sslKey" Git
configuration values with that of Git itself.

As part of this change, the "clone ClientCert with homedir certs" test
was added to our t/t-clone.sh test script.  This test sets the
"http.<url>.sslCert" and "http.<url>.sslKey" Git configuration options
to paths beginning with a tilde, in order to validate our support for
such paths.

As well, the following test, the "clone with flags" test, was updated
in the same commit to include a pair of commands which reset the
"http.<url>.sslCert" and "http.<url>.sslKey" Git configuration options
to their usual values; that is, to the values set by the setup() shell
function run at the start of each test file.  This makes sense, because
the goal was to ensure that in all cases, the tests following the new
"clone ClientCert with homedir certs" test had the expected values for
these configuration options.

However, resetting these values is not actually required, because
every test starts with the begin_test() shell function, which already
copies a fresh version of the global Git configuration file into the
directory used as the value of the HOME environment variable in each
test.  So we can simply remove the two "git config" calls at the start
of the "clone with flags" test, as the values they set are always the
same as the ones provided by our default test environment.