t/cmd/lfstest-gitserver.go: use uppercase hex salt
Due to the issue described in #5658, the "clone ClientCert" test in our
t/t-clone.sh test script does not actually run to completion, because
it exits early due to an improper check of the TRAVIS variable (which
is, itself, no longer used since we migrated our test suite to GitHub
Actions in PR #3808).

We will address this issue in a subsequent commit in this PR, at which
point the test will run fully, exposing a number of long-standing
problems with the test.  One of these occurs on Debian and Ubuntu Linux
systems where the libcurl library used by Git depends on a version of
the GnuTLS library older than v3.8.0, which at the moment, includes all
of our Debian build target platforms (Debian 10, 11, and 12).

Version 3.8.0 of GnuTLS contains a bug fix which allows it to parse and
accept DEK-Info PEM headers with lowercase hexadecimal strings.  Prior
to this change, it would reject these with an "invalid request" error:

  gnutls/gnutls@4604bbde14d2c6adb2af5315f9063ad65ab50aa6
  https://gitlab.com/gnutls/gnutls/-/merge_requests/1655

As it happens, the encrypted TLS certificate key generated by our
lfstest-gitserver test utility has its DEK-Info header formatted with
lowercase "a" through "f" letters, as this is what the EncodeToString()
function of the Go language "encoding/hex" package generates, and
that function is used by the (deprecated) EncryptPEMBlock() function
of the "crypto/x509" package which our test server calls to create the
encrypted key:

  https://github.com/golang/go/blob/7529d09a11496a77ccbffe245607fbd256200991/src/encoding/hex/hex.go#L17
  https://github.com/golang/go/blob/7529d09a11496a77ccbffe245607fbd256200991/src/crypto/x509/pem_decrypt.go#L228

When our "clone ClientCert" test runs on a Debian system, GnuTLS rejects
this key, which means Git in turn rejects it, and the test fails.
Our "clone ClientCert" test is the only one which validates the use
of the encrypted key, so it alone is affected by this problem, and because
the test always exits early at the moment (due to the issue described
in #5658), we have not noticed the failure until now.

We can work around the problem simply by rewriting the DEK-Info header
returned by the EncryptPEMBlock() function to use uppercase "A" through
"F" letters, which in turns allows GnuTLS to validate the salt, so
Git proceeds and our test succeeds.