t/cmd/git-credential-lfstest.go: do not echo lines
At present, our credential helper test program echoes a number of
identifying input lines back to the client, including the "host",
"protocol", and "path" lines (as well as certain "capability[]" lines
when those are appropriate).

However, common credential helper programs such as Git's own
git-credential-store and the git-credential-osxkeychain helper for
macOS generally avoid echoing such input lines back to the client,
with the exception of "capability[]" lines where the helper is
confirming which of the client's announced capabilities it supports.

This is roughly in line with the advice in Git's documentation,
which recommends against helpers overriding the values sent by clients,
other than for "username" and "password" fields.

  https://github.com/git/git/blob/4590f2e9412378c61eac95966709c78766d326ba/Documentation/gitcredentials.txt#L288-L290

Of course, just echoing the client's data back will cause no
functional difference in the client's behaviour, and is not a case
where the helper is overriding the data provided by the client with
different values.

Nevertheless, we can simplify our helper's logic and output, and
bring its operation into closer alignment with that of actual
credential helper programs, if we do not echo the "host", "protocol",
and "path" fields back to the client.  This also allows us to
exit early in the case where there are no fields to return other
than possibly some "capability[]" ones.