t/cmd/git-credential-lfstest.go: amend file lookup
Our credential helper test program looks for a credential record file in
the directory specified by the CREDSDIR environment variable, using the
values of the "host" and, if provided, "path" fields from the input lines
sent by a client.  At the moment, the correct construction of the paths
and names of the files which the program then tries to open depends on
the path in the CREDSDIR environment variable ending with a slash
character.

While our test suite always supplies a path with a trailing slash in
the CREDSDIR variable, the requirement to do this adds a small amount of
friction when debugging or testing our helper program.  By refactoring
the credsForHostAndPath() function in the program, we can both remove
the necessity for a trailing slash on the CREDSDIR path and clarify
the logic by which the function constructs file names and paths,
so we do that now.