t/testenv.sh: drop unused configuration variable
The LFS_CONFIG variable, which we set in our t/testenv.sh library,
is not referenced by any part of our current test suite.  The last
use of this variable was removed in 2015 in commit
1feee5837eea5a5e5d8e5a3ef387415ffe58ce87 of PR #335, when the
test suite was revised to create a private home directory for the
tests, so as to be able to set a standard global Git configuration
separate from that of the current user's configuration.  Hence we
can just remove the LFS_CONFIG variable now.