t/t-clone.sh: always add and commit .gitattributes
In commit cdec9f24fef979ed9e7fc6eab58ea91a8e6cf318 of PR #5699 we
updated two tests in our t/t-clone.sh test script to ensure they
add and commit the ".gitattributes" files created by "git lfs track"
before proceeding to generate commit data with Git LFS objects.
Per the commit description, this change was made so as to be explicit
about the order of expected Git operations, and to avoid relying
on the "git lfs clone" command to search for Git LFS pointer files
in the current Git working tree.

However, two other tests in the same test script have not run for
a long time, due to the bug described in #5658, where an improper
check of the TRAVIS variable causes the tests to exit early and
always declare success.  Hence, the need to make corresponding
changes to ensure these tests also explicitly add and commit
their ".gitattributes" files was not clear at the time when
commit cdec9f24fef979ed9e7fc6eab58ea91a8e6cf318 was written.

We will enable these skipped tests in a subsequent commit in this PR,
at which time they will fail unless we also adjust them to call
"git add" and "git commit" on their ".gitattributes" files, so we
do that now.

We also take the opportunity to make matching changes to a third
pair of tests, the "clone ClientCert with homedir certs" and "clone
with flags" tests, although these do run fully and successfully now.
While they do not depend on the ".gitattributes" file being committed
to their respective test repositories, aligning their code with that
of the other tests in the same test script may help avoid problems
in the future.