t/t-clone.sh: update progress meter message check
In commit db4e3b271d08da8110be0ee0ee091a2f9f212fbc of PR #2811 we
updated the progress meter text of the Git LFS client from "Git LFS"
to the more explicit "Uploading LFS objects" or "Downloading LFS
objects".  A large number of tests were changed in the same commit
because they check the progress meter output in logs they capture
from Git LFS operations.

However, the "cloneSSL" test in our t/t-clone.sh test script was not
updated because it has not actually run to completion for a long time,
as described in #5658, so it did not fail when the progress meter text
changed.  Due to a bug in the conditional used to check the TRAVIS
variable, the test instead always exits early and declares success.

We will address this problem in a subsequent commit in this PR,
at which point the test will fail unless we adjust it to check
for the same progress meter text as all our other tests, so we do
that now.

We also take the opportunity to add a similar check of the same
progress meter message to the "clone ClientCert" test.  This brings
our tests in the t/t-clone.sh script into slightly closer alignment
with each other, which should help ensure they stay consistent in
the future.  (Note that this test does not run to completion at the
moment either, for the same reason as the "cloneSSL" test.)