t/t-clone.sh: assert clean status when appropriate
In commit caaa0f91fad1afccc7c40197882160aea8b06d3c of PR #2647 a
number of tests of the "git lfs clone" and "git lfs pull" commands
were enhanced so as to check that after those commands are invoked,
the "git status" command returns a "working tree clean" message.
To perform these checks, a call to our assert_clean_status()
shell function was added to the tests.

In the case of the "cloneSSL" test, an assert_clean_status() call
was added, but left commented out; it was then uncommented in commit
e0eede1602b09f1d3e81006ff7b30475f449b576 of the same PR.  Unfortunately,
the call is made when the current working directory has not yet been
changed to that of the newly cloned repository's working tree, and
so should fail as it stands now.

However, as described in #5658, the "cloneSSL" test and the "clone
ClientCert" tests in our t/t-clone.sh test script do not actually
run to completion, a consequence of an improper check of the TRAVIS
variable (which is no longer used since we migrated our test suite to
GitHub Actions in PR #3808).  Instead, the tests exit early and
always declare success.

We will address this problem in a subsequent commit in this PR, and when
we do so, the "cloneSSL" test will fail because the misplaced assertion.
Therefore we move the assertion into the block of checks performed
after a "pushd" shell command changes the current directory to that
of the new clone's working tree, which will allow the test to pass
when we resolve the bug with the TRAVIS variable check.

We also take the opportunity to add assert_clean_status() calls to a
number of other tests in the t/t-clone.sh test script, so the tests are
all performing similar sets of checks.  This will help us keep the
tests in that script more consistent with each other in the future.