git/git.go: report unexpected tag as error
In a previous commit in this PR we introduced the ignoreTags argument
to the RemoteRefs() function in our "git" package, so that we could
indicate to the function whether or not it should pass the --tags option
to the "git ls-remote" command.  When the ignoreTags flag is set, we do
not pass the --tags option to the command, only the --heads option.

As a result, the "git ls-remote" command should never return tag refs
in this case, so we can now add a check and return an error from the
RemoteRefs() function if a tag ref is found in the set of refs returned
by "git ls-remote" when ignoreTags is "true".

We also remove some excess whitespace to satisfy the "go fmt" linter.