lfs/gitscanner_remotes.go: assume only branches
In a previous commit in this PR we revised the RemoteRefs() function
in our "git" package to take an additional ignoreTags argument, and
updated the callers of this function appropriately.  In the
calcSkippedRefs() function of our "lfs" package we now pass a "true"
value to the RemoteRefs() function, which guarantees it will only
return branch refs.

The calcSkippedRefs() function compares this set of branch refs
against the refs returned from the "git" package's CachedRemoteRefs()
function, which also consists exclusively of branch refs.  Any refs
which are in both sets are then returned as a list of refs to be
ignored.  In particular, in subsequent calls to the "git rev-list"
command, any commits reachable from the commits referenced by the
locally cached versions of these ignorable refs will be excluded.

As the RemoteRefs() function will no longer return any tags, we
can slightly simplify the calcSkippedRefs() function, since it does
not need to check for tag refs any longer.  We also rename the
actualRemoteBranchRefs variable to actualRemoteRefsSet, and revise
and expand some inline comments to clarify the code's expectations
and intent.