t/t-push.sh: add push test with multiple tags
In previous commits in this PR we have adjusted the arguments we use
when executing the "git ls-remote" command so as to limit the refs it
returns to include only branches, in cases where we do not need to
process any tag refs.

Specifically, the "git lfs push" and "git lfs pre-push" commands invoke
the NewGitScannerForPush() function in our "lfs" package, which uses
the calcSkippedRefs() function in the same package to find the set
of remote branch refs that are also cached locally.  Any commits
reachable from the commits referenced by the locally cached versions
of these refs will then be excluded from the set of commits returned
by the "git rev-list" command when that is subsequently called by
the revListShas() function and the methods of the RevListScanner
structure in our "git" package.  No tag refs are considered during
these steps, so can avoid retrieving them with the "git ls-remote"
command.

Earlier, in commit aab142ddb984794d15e0f2b861442e8e55951969 of PR #4102,
we added the "pre-push with force-pushed ref" test to our t/t-pre-push.sh
test suite, which checks that the "git lfs pre-push" command handles
the condition where a Git commit is referenced by a remote ref, and
is not present locally, but also happens to be one that can be ignored
because it has been replaced in the local version of the ref by a
force-push.  (The test runs the "git lfs pre-push" command indirectly,
as it is executed by the "git push" command.)

Fortunately, this test suffices to validate the changes we have made
in this PR also, at least for the "git lfs pre-push" command, because the
test creates and pushes a tag, and then updates and pushes the tag again.
As a result, it runs the "git lfs pre-push" command in a context where
a tag ref exists on the Git remote.  Prior to our changes in this PR,
the "git ls-remote" command (as performed by the RemoteRefs() function
in the "git" package, when called by the calcSkippedRefs() function)
would return the tag ref as well as the "main" branch ref.  With this
PR's changes, the "git ls-remote" command now only returns the "main"
branch ref under this test's conditions, so we have an effective test
of the "git lfs pre-push" command with a tag ref on a Git remote.

However, we do not have a test of the "git lfs push" command with a
tag ref on the Git remote, so we add such a test now.  This test is
not strictly necessary, but ensures we run the "git lfs push" command
in a context where a tag has been created and pushed to the Git remote.
Without this PR's other changes, the tag ref would then be returned by
"git ls-remote" when that was performed as part of the "git lfs push"
command's operation.  With this PR's changes, though, the tag ref will
not be returned by "git ls-remote".  The test will succeed in either
case because the tag ref is ultimately immaterial in the determination
of the refs returned by the calcSkippedRefs() function.